✨ feat: 알림 누르면 오늘의 뉴스뷰로 이동
