+++ NewsHabit/Sources/AppDelegate.swift
@@ -6,7 +6,6 @@
 //
 
 import UIKit
-import UserNotifications
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -17,27 +16,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             UserDefaultsManager.lastMonth = Date().toMonthString()
             UserDefaultsManager.monthlyAllRead = []
         }
-        // 알림 센터의 delegate 설정
-        UNUserNotificationCenter.current().delegate = self
-        NotificationCenterManager.shared.requestAuthorization { isAuthorized, error in
-            UserDefaultsManager.isNotificationOn = isAuthorized
-            if isAuthorized {
-                if let notificationTime = UserDefaultsManager.notificationTime.toTimeAsDate() {
-                    NotificationCenterManager.shared.addNotification(for: notificationTime)
-                }
-            }
-        }
         return true
-    }
-    
-}
-
-extension AppDelegate: UNUserNotificationCenterDelegate {
-    
-    // 앱이 포그라운드 상태일 때 알림이 도착하면 호출
-    func userNotificationCenter(_ center: UNUserNotificationCenter, willPresent notification: UNNotification, withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
-        // 알림 배너, 소리 등을 표시하도록 설정
-        completionHandler([.banner, .badge, .list, .sound])
     }
     
 }
